requirements.txt
pygame>=2.5.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0